+++ index.html
@@ -230,9 +230,6 @@
     </div>
   </template>
 
-<script src="./js/main.js"></script>
-<script src="./js/data.js" type="module"></script>
-<script src="./js/util.js" type="module"></script>
-<script src="./js/card.js" type="module"></script>
+<script src="./js/main.js" type="module"></script>
 </body>
 </html>
